Fix compile and warnings on BSD.

+++ script.c
@@ -383,16 +383,16 @@ make_env(const struct interface *ifp, const char *reason, char ***argv)
 		else	
 			pfx = NULL;
 		if (pfx) {
-			int pfx_len;
+			size_t pfx_len;
 			ssize_t psl;
 
 			pfx_len = strlen(pfx);
 			psl = print_string(NULL, 0,
 			    (const uint8_t *)ifp->ssid, ifp->ssid_len);
 			if (psl != -1) {
-				EMALLOC(elen, pfx_len + psl + 1);
+				EMALLOC(elen, pfx_len + (size_t)psl + 1);
 				memcpy(env[elen], pfx, pfx_len);
-				print_string(env[elen] + pfx_len, psl,
+				print_string(env[elen] + pfx_len, (size_t)psl,
 				    (const uint8_t *)ifp->ssid, ifp->ssid_len);
 				elen++;
 			}
